move tests from verification module to macros playground module;
