+++ sigma-dsl-compiler-macros-playground/src/test/scala/sigmastate/verification/test/MiscGenerators.scala
