+++ sigma-dsl-compiler-macros-playground/src/test/scala/sigmastate/verification/test/VerifiedTypeConvertersTest.scala
