+++ stainless.conf
@@ -3,7 +3,6 @@
 verification = true
 termination = true
 strict-arithmetic = false
-functions = [DummyContractVerification._, AssetsAtomicExchangeVerification._]
 timeout = 10
 print-ids = false
 # debug = [verification]
